requirements.txt
apiview==1.3.25
appdirs==1.4.4
apturl==0.5.2
attrs==20.3.0
autopep8==1.5.5
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Bottleneck==1.2.1
Brlapi==0.8.2
caer==2.0.8
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.4
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==2.2.24
django-cors-headers==3.11.0
django-heroku==0.3.1
django-import-export==2.7.1
djangorestframework==3.13.1
djlint==0.7.6
duplicity==0.8.20
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.11.3
iniconfig==0.0.0
jdcal==1.0
jeepney==0.7.1
Jinja2==2.11.3
keyring==23.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
llvmlite==0.35.0
lockfile==0.12.2
louis==3.18.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.4
monotonic==1.5
more-itertools==4.2.0
mypy==0.942
mypy-extensions==0.4.3
mysqlclient==2.1.0
natsort==7.1.0
netifaces==0.10.9
numba==0.52.0
numexpr==2.7.2
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-contrib-python==4.5.5.64
openpyxl==3.0.3
packaging==20.9
pandas==1.1.5
paramiko==2.7.2
pathspec==0.9.0
pep8==1.7.1
pexpect==4.8.0
Pillow==8.1.2
pipenv==11.9.0
pluggy==0.13.0
protobuf==3.12.4
psycopg2==2.9.3
py==1.10.0
pycairo==1.16.2
pycodestyle==2.6.0
pycups==2.0.1
Pygments==2.7.1
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.0.2
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
pytz==2021.1
pyxdg==0.27
PyYAML==6.0
regex==2022.3.15
reportlab==3.5.66
requests==2.25.1
scipy==1.6.0
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
systemd-python==234
tables==3.6.1
tablib==3.2.0
toml==0.10.1
tomli==2.0.1
tomlkit==0.9.2
tqdm==4.63.1
ttystatus==0.38
typecatcher==0.3
typing-extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.4.6+ds
virtualenv-clone==0.3.0
wadllib==1.3.5
webencodings==0.5.1
whitenoise==6.0.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==1.0.0
